+++ examples/rainmaker_light/main/app_rainmaker.h
@@ -47,7 +47,8 @@ esp_err_t app_rainmaker_start(void);
  * @return ESP_OK on success.
  * @return error in case of failure.
  */
-esp_err_t app_rainmaker_attribute_update(int endpoint_id, int cluster_id, int attribute_id, esp_matter_attr_val_t *val);
+esp_err_t app_rainmaker_attribute_update(uint16_t endpoint_id, uint32_t cluster_id, uint32_t attribute_id,
+                                         esp_matter_attr_val_t *val);
 
 #ifdef __cplusplus
 }
